fixing betty

+++ 0x08-palindrome_integer/0-is_palindrome.c
@@ -25,22 +25,22 @@ int len_n(unsigned long n)
  */
 int is_palindrome(unsigned long n)
 {
-	unsigned store[20];
+	unsigned long store[20];
 	int num, i;
 
 	num = len_n(n);
 
-	for(i = 0; i < num / 2; i++)
+	for (i = 0; i < num / 2; i++)
 	{
 		store[i] = n % 10;
 		n = n / 10;
 	}
 	i--;
 
-	if(num % 2 == 1) 
+	if (num % 2 == 1)
 		n = n / 10;
-    
-	for(; i >= 0; i--)
+
+	for (; i >= 0; i--)
 	{
 		if (n % 10 != store[i])
 			return (0);
